Module specs::data

Reexports for types implementing SystemData.

Examples

These can be used in a System implementation

use specs::prelude::*;

struct MySys;

impl<'a> System<'a> for MySys {
    type SystemData = (Entities<'a>, FetchMut<'a, MyRes>, WriteStorage<'a, MyComp>);

    fn run(&mut self, data: Self::SystemData) {
        // ..

    }
}

Structs

Fetch
Return value of Resources::fetch.

FetchId
Return value of Resources::fetch_id.

FetchIdMut
Return value of Resources::fetch_id_mut.

FetchMut
Return value of Resources::fetch_mut.

Storage
A wrapper around the masked storage and the generations vector. Can be used for safe lookup of components, insertions and removes. This is what World::read/write fetches for the user.

Traits

SystemData
A struct implementing system data indicates that it bundles some resources which are required for the execution.

Type Definitions

Entities
A wrapper for a fetched Entities resource. Note that this is just Fetch<Entities>, so you can easily use it in your system:

ReadStorage
A storage with read access.

WriteStorage
A storage with read and write access.