Struct specs::data::Storage

pub struct Storage<'e, T, D> { /* fields omitted */ }

A wrapper around the masked storage and the generations vector. Can be used for safe lookup of components, insertions and removes. This is what World::read/write fetches for the user.

Methods

impl<'e, T, D> Storage<'e, T, D> where
T: Component,
D: Deref<Target = MaskedStorage<T>>,

fn check(&self) -> CheckStorage<'e, T, D>

Returns a struct that can iterate over the entities that have it but does not return the contents of the storage.

Useful if you want to check if an entity has a component and then possibly get the component later on in the loop.

fn get_unchecked<'a>(&'a self, entry: &'a Entry<'a, 'e, T, D>) -> &'a T

Reads the data associated with the entry.

Entrys are returned from a CheckStorage to remove unnecessary checks.

Panics

Panics if the entry was retrieved from another storage.

impl<'e, T, D> Storage<'e, T, D> where
T: Component,
D: DerefMut<Target = MaskedStorage<T>>,

fn get_mut_unchecked<'a>(
&'a mut self,
entry: &'a mut Entry<'a, 'e, T, D>
) -> &'a mut T

Tries to mutate the data associated with an entry.

Entrys are returned from a CheckStorage to remove unnecessary checks.

impl<'e, T, D> Storage<'e, T, D>

fn new(entities: Fetch<'e, Entities>, data: D) -> Storage<'e, T, D>

Create a new Storage

impl<'e, T, D> Storage<'e, T, D> where
T: Component,
D: Deref<Target = MaskedStorage<T>>,

fn get(&self, e: Entity) -> Option<&T>

Tries to read the data associated with an Entity.

impl<'e, T, D> Storage<'e, T, D> where
T: Component,
D: DerefMut<Target = MaskedStorage<T>>,

fn get_mut(&mut self, e: Entity) -> Option<&mut T>

Tries to mutate the data associated with an Entity.

fn insert(&mut self, e: Entity, v: T) -> InsertResult<T>

Inserts new data for a given Entity. Returns the result of the operation as a InsertResult<T>

fn remove(&mut self, e: Entity) -> Option<T>

Removes the data associated with an Entity.

fn clear(&mut self)

Clears the contents of the storage.