Enum specs::InsertResult

pub enum InsertResult<T> {
    Inserted,
    Updated(T),
    EntityIsDead(T),
}

The status of an insert()ion into a storage.

Variants
Inserted

The value was inserted and there was no value before

Updated(T)

The value was updated an already inserted value the value returned is the old value

EntityIsDead(T)

The value failed to insert because the entity was invalid