Module specs::join

Joining of components for iteration over entities with specific components.

Structs

JoinIter
JoinIter is an Iterator over a group of Storages.

JoinParIter
JoinParIter is a ParallelIterator over a group of Storages.

MaybeJoin
A Join-able structure that yields all indices, returning None for all missing elements and Some(T) for found elements.

Traits

BitAnd
BitAnd is a helper method to & bitsets together resulting in a tree.

Join
The purpose of the Join trait is to provide a way to access multiple storages at the same time with the merged bit set.

ParJoin
The purpose of the ParJoin trait is to provide a way to access multiple storages in parallel at the same time with the merged bit set.