Struct specs::join::MaybeJoin

pub struct MaybeJoin<J: Join>(pub J);

A Join-able structure that yields all indices, returning None for all missing elements and Some(T) for found elements.

For usage see Join::maybe().

WARNING: Do not have a join of only MaybeJoins. Otherwise the join will iterate over every single index of the bitset. If you want a join with all MaybeJoins, add an EntitiesRes to the join as well to bound the join to all entities that are alive.