Trait specs::join::ParJoin

pub unsafe trait ParJoin: Join {
    fn par_join(self) -> JoinParIter<Self>
    where
        Self: Sized,
    { ... }
}

The purpose of the ParJoin trait is to provide a way to access multiple storages in parallel at the same time with the merged bit set.

Safety

The implementation of ParallelIterator for ParJoin makes multiple assumptions on the structure of Self. In particular, <Self as Join>::get must be callable from multiple threads, simultaneously, without mutating values not exclusively associated with id.

Provided methods

fn par_join(self) -> JoinParIter<Self> where
Self: Sized,

Create a joined parallel iterator over the contents.

Loading content...