Trait specs::prelude::RunNow

pub trait RunNow<'a> {
    fn run_now(&mut self, world: &'a World);

    fn setup(&mut self, world: &mut World);

    fn dispose(self: Box<Self>, world: &mut World) { ... }
}

Trait for fetching data and running systems. Automatically implemented for systems.

Required methods

fn run_now(&mut self, world: &'a World)

Runs the system now.

Panics

Panics if the system tries to fetch resources which are borrowed in an incompatible way already (tries to read from a resource which is already written to or tries to write to a resource which is read from).

fn setup(&mut self, world: &mut World)

Sets up World for a later call to run_now.

Loading content...
Provided methods

fn dispose(self: Box<Self>, world: &mut World)

Performs clean up that requires resources from the World. This commonly removes components from world which depend on external resources.

Loading content...