Type Definition specs::prelude::WriteExpect

type WriteExpect<'a, T> = Write<'a, T, PanicHandler>;

Allows to fetch a resource in a system mutably. This will panic if the resource does not exist. Usage of Write or Option<Write> is therefore recommended.

If the nightly feature of shred is enabled, this will print the type of the resource in case of a panic. That can help for debugging.