Struct specs::saveload::WorldSerialize

pub struct WorldSerialize<'a, M: Marker, E, T: Components<M::Identifier, E>> { /* fields omitted */ }

This type implements Serialize so that it may be used in generic environment where Serialize implementation is expected. It may be constructed manually with WorldSerialize::new. Or fetched from System as SystemData. Serializes components in tuple T with marker M.

Methods

impl<'a, X, Z, A, B, C, D, E, F, G, H, I, J, K, L, M, N, O> WorldSerialize<'a, X, Z, (A, B, C, D, E, F, G, H, I, J, K, L, M, N, O)> where
X: Marker,
A: SaveLoadComponent<X::Identifier>,
Z: From<A::Error>,
B: SaveLoadComponent<X::Identifier>,
Z: From<B::Error>,
C: SaveLoadComponent<X::Identifier>,
Z: From<C::Error>,
D: SaveLoadComponent<X::Identifier>,
Z: From<D::Error>,
E: SaveLoadComponent<X::Identifier>,
Z: From<E::Error>,
F: SaveLoadComponent<X::Identifier>,
Z: From<F::Error>,
G: SaveLoadComponent<X::Identifier>,
Z: From<G::Error>,
H: SaveLoadComponent<X::Identifier>,
Z: From<H::Error>,
I: SaveLoadComponent<X::Identifier>,
Z: From<I::Error>,
J: SaveLoadComponent<X::Identifier>,
Z: From<J::Error>,
K: SaveLoadComponent<X::Identifier>,
Z: From<K::Error>,
L: SaveLoadComponent<X::Identifier>,
Z: From<L::Error>,
M: SaveLoadComponent<X::Identifier>,
Z: From<M::Error>,
N: SaveLoadComponent<X::Identifier>,
Z: From<N::Error>,
O: SaveLoadComponent<X::Identifier>,
Z: From<O::Error>,

pub fn new(
entities: Entities<'a>,
markers: ReadStorage<'a, X>,
A: ReadStorage<'a, A>,
B: ReadStorage<'a, B>,
C: ReadStorage<'a, C>,
D: ReadStorage<'a, D>,
E: ReadStorage<'a, E>,
F: ReadStorage<'a, F>,
G: ReadStorage<'a, G>,
H: ReadStorage<'a, H>,
I: ReadStorage<'a, I>,
J: ReadStorage<'a, J>,
K: ReadStorage<'a, K>,
L: ReadStorage<'a, L>,
M: ReadStorage<'a, M>,
N: ReadStorage<'a, N>,
O: ReadStorage<'a, O>
) -> Self

Create serializable structure from storages

impl<'a, X, Z, B, C, D, E, F, G, H, I, J, K, L, M, N, O> WorldSerialize<'a, X, Z, (B, C, D, E, F, G, H, I, J, K, L, M, N, O)> where
X: Marker,
B: SaveLoadComponent<X::Identifier>,
Z: From<B::Error>,
C: SaveLoadComponent<X::Identifier>,
Z: From<C::Error>,
D: SaveLoadComponent<X::Identifier>,
Z: From<D::Error>,
E: SaveLoadComponent<X::Identifier>,
Z: From<E::Error>,
F: SaveLoadComponent<X::Identifier>,
Z: From<F::Error>,
G: SaveLoadComponent<X::Identifier>,
Z: From<G::Error>,
H: SaveLoadComponent<X::Identifier>,
Z: From<H::Error>,
I: SaveLoadComponent<X::Identifier>,
Z: From<I::Error>,
J: SaveLoadComponent<X::Identifier>,
Z: From<J::Error>,
K: SaveLoadComponent<X::Identifier>,
Z: From<K::Error>,
L: SaveLoadComponent<X::Identifier>,
Z: From<L::Error>,
M: SaveLoadComponent<X::Identifier>,
Z: From<M::Error>,
N: SaveLoadComponent<X::Identifier>,
Z: From<N::Error>,
O: SaveLoadComponent<X::Identifier>,
Z: From<O::Error>,

pub fn new(
entities: Entities<'a>,
markers: ReadStorage<'a, X>,
B: ReadStorage<'a, B>,
C: ReadStorage<'a, C>,
D: ReadStorage<'a, D>,
E: ReadStorage<'a, E>,
F: ReadStorage<'a, F>,
G: ReadStorage<'a, G>,
H: ReadStorage<'a, H>,
I: ReadStorage<'a, I>,
J: ReadStorage<'a, J>,
K: ReadStorage<'a, K>,
L: ReadStorage<'a, L>,
M: ReadStorage<'a, M>,
N: ReadStorage<'a, N>,
O: ReadStorage<'a, O>
) -> Self

Create serializable structure from storages

impl<'a, X, Z, C, D, E, F, G, H, I, J, K, L, M, N, O> WorldSerialize<'a, X, Z, (C, D, E, F, G, H, I, J, K, L, M, N, O)> where
X: Marker,
C: SaveLoadComponent<X::Identifier>,
Z: From<C::Error>,
D: SaveLoadComponent<X::Identifier>,
Z: From<D::Error>,
E: SaveLoadComponent<X::Identifier>,
Z: From<E::Error>,
F: SaveLoadComponent<X::Identifier>,
Z: From<F::Error>,
G: SaveLoadComponent<X::Identifier>,
Z: From<G::Error>,
H: SaveLoadComponent<X::Identifier>,
Z: From<H::Error>,
I: SaveLoadComponent<X::Identifier>,
Z: From<I::Error>,
J: SaveLoadComponent<X::Identifier>,
Z: From<J::Error>,
K: SaveLoadComponent<X::Identifier>,
Z: From<K::Error>,
L: SaveLoadComponent<X::Identifier>,
Z: From<L::Error>,
M: SaveLoadComponent<X::Identifier>,
Z: From<M::Error>,
N: SaveLoadComponent<X::Identifier>,
Z: From<N::Error>,
O: SaveLoadComponent<X::Identifier>,
Z: From<O::Error>,

pub fn new(
entities: Entities<'a>,
markers: ReadStorage<'a, X>,
C: ReadStorage<'a, C>,
D: ReadStorage<'a, D>,
E: ReadStorage<'a, E>,
F: ReadStorage<'a, F>,
G: ReadStorage<'a, G>,
H: ReadStorage<'a, H>,
I: ReadStorage<'a, I>,
J: ReadStorage<'a, J>,
K: ReadStorage<'a, K>,
L: ReadStorage<'a, L>,
M: ReadStorage<'a, M>,
N: ReadStorage<'a, N>,
O: ReadStorage<'a, O>
) -> Self

Create serializable structure from storages

impl<'a, X, Z, D, E, F, G, H, I, J, K, L, M, N, O> WorldSerialize<'a, X, Z, (D, E, F, G, H, I, J, K, L, M, N, O)> where
X: Marker,
D: SaveLoadComponent<X::Identifier>,
Z: From<D::Error>,
E: SaveLoadComponent<X::Identifier>,
Z: From<E::Error>,
F: SaveLoadComponent<X::Identifier>,
Z: From<F::Error>,
G: SaveLoadComponent<X::Identifier>,
Z: From<G::Error>,
H: SaveLoadComponent<X::Identifier>,
Z: From<H::Error>,
I: SaveLoadComponent<X::Identifier>,
Z: From<I::Error>,
J: SaveLoadComponent<X::Identifier>,
Z: From<J::Error>,
K: SaveLoadComponent<X::Identifier>,
Z: From<K::Error>,
L: SaveLoadComponent<X::Identifier>,
Z: From<L::Error>,
M: SaveLoadComponent<X::Identifier>,
Z: From<M::Error>,
N: SaveLoadComponent<X::Identifier>,
Z: From<N::Error>,
O: SaveLoadComponent<X::Identifier>,
Z: From<O::Error>,

pub fn new(
entities: Entities<'a>,
markers: ReadStorage<'a, X>,
D: ReadStorage<'a, D>,
E: ReadStorage<'a, E>,
F: ReadStorage<'a, F>,
G: ReadStorage<'a, G>,
H: ReadStorage<'a, H>,
I: ReadStorage<'a, I>,
J: ReadStorage<'a, J>,
K: ReadStorage<'a, K>,
L: ReadStorage<'a, L>,
M: ReadStorage<'a, M>,
N: ReadStorage<'a, N>,
O: ReadStorage<'a, O>
) -> Self

Create serializable structure from storages

impl<'a, X, Z, E, F, G, H, I, J, K, L, M, N, O> WorldSerialize<'a, X, Z, (E, F, G, H, I, J, K, L, M, N, O)> where
X: Marker,
E: SaveLoadComponent<X::Identifier>,
Z: From<E::Error>,
F: SaveLoadComponent<X::Identifier>,
Z: From<F::Error>,
G: SaveLoadComponent<X::Identifier>,
Z: From<G::Error>,
H: SaveLoadComponent<X::Identifier>,
Z: From<H::Error>,
I: SaveLoadComponent<X::Identifier>,
Z: From<I::Error>,
J: SaveLoadComponent<X::Identifier>,
Z: From<J::Error>,
K: SaveLoadComponent<X::Identifier>,
Z: From<K::Error>,
L: SaveLoadComponent<X::Identifier>,
Z: From<L::Error>,
M: SaveLoadComponent<X::Identifier>,
Z: From<M::Error>,
N: SaveLoadComponent<X::Identifier>,
Z: From<N::Error>,
O: SaveLoadComponent<X::Identifier>,
Z: From<O::Error>,

pub fn new(
entities: Entities<'a>,
markers: ReadStorage<'a, X>,
E: ReadStorage<'a, E>,
F: ReadStorage<'a, F>,
G: ReadStorage<'a, G>,
H: ReadStorage<'a, H>,
I: ReadStorage<'a, I>,
J: ReadStorage<'a, J>,
K: ReadStorage<'a, K>,
L: ReadStorage<'a, L>,
M: ReadStorage<'a, M>,
N: ReadStorage<'a, N>,
O: ReadStorage<'a, O>
) -> Self

Create serializable structure from storages

impl<'a, X, Z, F, G, H, I, J, K, L, M, N, O> WorldSerialize<'a, X, Z, (F, G, H, I, J, K, L, M, N, O)> where
X: Marker,
F: SaveLoadComponent<X::Identifier>,
Z: From<F::Error>,
G: SaveLoadComponent<X::Identifier>,
Z: From<G::Error>,
H: SaveLoadComponent<X::Identifier>,
Z: From<H::Error>,
I: SaveLoadComponent<X::Identifier>,
Z: From<I::Error>,
J: SaveLoadComponent<X::Identifier>,
Z: From<J::Error>,
K: SaveLoadComponent<X::Identifier>,
Z: From<K::Error>,
L: SaveLoadComponent<X::Identifier>,
Z: From<L::Error>,
M: SaveLoadComponent<X::Identifier>,
Z: From<M::Error>,
N: SaveLoadComponent<X::Identifier>,
Z: From<N::Error>,
O: SaveLoadComponent<X::Identifier>,
Z: From<O::Error>,

pub fn new(
entities: Entities<'a>,
markers: ReadStorage<'a, X>,
F: ReadStorage<'a, F>,
G: ReadStorage<'a, G>,
H: ReadStorage<'a, H>,
I: ReadStorage<'a, I>,
J: ReadStorage<'a, J>,
K: ReadStorage<'a, K>,
L: ReadStorage<'a, L>,
M: ReadStorage<'a, M>,
N: ReadStorage<'a, N>,
O: ReadStorage<'a, O>
) -> Self

Create serializable structure from storages

impl<'a, X, Z, G, H, I, J, K, L, M, N, O> WorldSerialize<'a, X, Z, (G, H, I, J, K, L, M, N, O)> where
X: Marker,
G: SaveLoadComponent<X::Identifier>,
Z: From<G::Error>,
H: SaveLoadComponent<X::Identifier>,
Z: From<H::Error>,
I: SaveLoadComponent<X::Identifier>,
Z: From<I::Error>,
J: SaveLoadComponent<X::Identifier>,
Z: From<J::Error>,
K: SaveLoadComponent<X::Identifier>,
Z: From<K::Error>,
L: SaveLoadComponent<X::Identifier>,
Z: From<L::Error>,
M: SaveLoadComponent<X::Identifier>,
Z: From<M::Error>,
N: SaveLoadComponent<X::Identifier>,
Z: From<N::Error>,
O: SaveLoadComponent<X::Identifier>,
Z: From<O::Error>,

pub fn new(
entities: Entities<'a>,
markers: ReadStorage<'a, X>,
G: ReadStorage<'a, G>,
H: ReadStorage<'a, H>,
I: ReadStorage<'a, I>,
J: ReadStorage<'a, J>,
K: ReadStorage<'a, K>,
L: ReadStorage<'a, L>,
M: ReadStorage<'a, M>,
N: ReadStorage<'a, N>,
O: ReadStorage<'a, O>
) -> Self

Create serializable structure from storages

impl<'a, X, Z, H, I, J, K, L, M, N, O> WorldSerialize<'a, X, Z, (H, I, J, K, L, M, N, O)> where
X: Marker,
H: SaveLoadComponent<X::Identifier>,
Z: From<H::Error>,
I: SaveLoadComponent<X::Identifier>,
Z: From<I::Error>,
J: SaveLoadComponent<X::Identifier>,
Z: From<J::Error>,
K: SaveLoadComponent<X::Identifier>,
Z: From<K::Error>,
L: SaveLoadComponent<X::Identifier>,
Z: From<L::Error>,
M: SaveLoadComponent<X::Identifier>,
Z: From<M::Error>,
N: SaveLoadComponent<X::Identifier>,
Z: From<N::Error>,
O: SaveLoadComponent<X::Identifier>,
Z: From<O::Error>,

pub fn new(
entities: Entities<'a>,
markers: ReadStorage<'a, X>,
H: ReadStorage<'a, H>,
I: ReadStorage<'a, I>,
J: ReadStorage<'a, J>,
K: ReadStorage<'a, K>,
L: ReadStorage<'a, L>,
M: ReadStorage<'a, M>,
N: ReadStorage<'a, N>,
O: ReadStorage<'a, O>
) -> Self

Create serializable structure from storages

impl<'a, X, Z, I, J, K, L, M, N, O> WorldSerialize<'a, X, Z, (I, J, K, L, M, N, O)> where
X: Marker,
I: SaveLoadComponent<X::Identifier>,
Z: From<I::Error>,
J: SaveLoadComponent<X::Identifier>,
Z: From<J::Error>,
K: SaveLoadComponent<X::Identifier>,
Z: From<K::Error>,
L: SaveLoadComponent<X::Identifier>,
Z: From<L::Error>,
M: SaveLoadComponent<X::Identifier>,
Z: From<M::Error>,
N: SaveLoadComponent<X::Identifier>,
Z: From<N::Error>,
O: SaveLoadComponent<X::Identifier>,
Z: From<O::Error>,

pub fn new(
entities: Entities<'a>,
markers: ReadStorage<'a, X>,
I: ReadStorage<'a, I>,
J: ReadStorage<'a, J>,
K: ReadStorage<'a, K>,
L: ReadStorage<'a, L>,
M: ReadStorage<'a, M>,
N: ReadStorage<'a, N>,
O: ReadStorage<'a, O>
) -> Self

Create serializable structure from storages

impl<'a, X, Z, J, K, L, M, N, O> WorldSerialize<'a, X, Z, (J, K, L, M, N, O)> where
X: Marker,
J: SaveLoadComponent<X::Identifier>,
Z: From<J::Error>,
K: SaveLoadComponent<X::Identifier>,
Z: From<K::Error>,
L: SaveLoadComponent<X::Identifier>,
Z: From<L::Error>,
M: SaveLoadComponent<X::Identifier>,
Z: From<M::Error>,
N: SaveLoadComponent<X::Identifier>,
Z: From<N::Error>,
O: SaveLoadComponent<X::Identifier>,
Z: From<O::Error>,

pub fn new(
entities: Entities<'a>,
markers: ReadStorage<'a, X>,
J: ReadStorage<'a, J>,
K: ReadStorage<'a, K>,
L: ReadStorage<'a, L>,
M: ReadStorage<'a, M>,
N: ReadStorage<'a, N>,
O: ReadStorage<'a, O>
) -> Self

Create serializable structure from storages

impl<'a, X, Z, K, L, M, N, O> WorldSerialize<'a, X, Z, (K, L, M, N, O)> where
X: Marker,
K: SaveLoadComponent<X::Identifier>,
Z: From<K::Error>,
L: SaveLoadComponent<X::Identifier>,
Z: From<L::Error>,
M: SaveLoadComponent<X::Identifier>,
Z: From<M::Error>,
N: SaveLoadComponent<X::Identifier>,
Z: From<N::Error>,
O: SaveLoadComponent<X::Identifier>,
Z: From<O::Error>,

pub fn new(
entities: Entities<'a>,
markers: ReadStorage<'a, X>,
K: ReadStorage<'a, K>,
L: ReadStorage<'a, L>,
M: ReadStorage<'a, M>,
N: ReadStorage<'a, N>,
O: ReadStorage<'a, O>
) -> Self

Create serializable structure from storages

impl<'a, X, Z, L, M, N, O> WorldSerialize<'a, X, Z, (L, M, N, O)> where
X: Marker,
L: SaveLoadComponent<X::Identifier>,
Z: From<L::Error>,
M: SaveLoadComponent<X::Identifier>,
Z: From<M::Error>,
N: SaveLoadComponent<X::Identifier>,
Z: From<N::Error>,
O: SaveLoadComponent<X::Identifier>,
Z: From<O::Error>,

pub fn new(
entities: Entities<'a>,
markers: ReadStorage<'a, X>,
L: ReadStorage<'a, L>,
M: ReadStorage<'a, M>,
N: ReadStorage<'a, N>,
O: ReadStorage<'a, O>
) -> Self

Create serializable structure from storages

impl<'a, X, Z, M, N, O> WorldSerialize<'a, X, Z, (M, N, O)> where
X: Marker,
M: SaveLoadComponent<X::Identifier>,
Z: From<M::Error>,
N: SaveLoadComponent<X::Identifier>,
Z: From<N::Error>,
O: SaveLoadComponent<X::Identifier>,
Z: From<O::Error>,

pub fn new(
entities: Entities<'a>,
markers: ReadStorage<'a, X>,
M: ReadStorage<'a, M>,
N: ReadStorage<'a, N>,
O: ReadStorage<'a, O>
) -> Self

Create serializable structure from storages

impl<'a, X, Z, N, O> WorldSerialize<'a, X, Z, (N, O)> where
X: Marker,
N: SaveLoadComponent<X::Identifier>,
Z: From<N::Error>,
O: SaveLoadComponent<X::Identifier>,
Z: From<O::Error>,

pub fn new(
entities: Entities<'a>,
markers: ReadStorage<'a, X>,
N: ReadStorage<'a, N>,
O: ReadStorage<'a, O>
) -> Self

Create serializable structure from storages

impl<'a, X, Z, O> WorldSerialize<'a, X, Z, (O,)> where
X: Marker,
O: SaveLoadComponent<X::Identifier>,
Z: From<O::Error>,

pub fn new(
entities: Entities<'a>,
markers: ReadStorage<'a, X>,
O: ReadStorage<'a, O>
) -> Self

Create serializable structure from storages

impl<'a, X, Z> WorldSerialize<'a, X, Z, ()> where
X: Marker,

pub fn new(entities: Entities<'a>, markers: ReadStorage<'a, X>) -> Self

Create serializable structure from storages

impl<'a, M, E, T> WorldSerialize<'a, M, E, T> where
M: Marker,
T: Components<M::Identifier, E>,

pub fn remove_serialized(&mut self)

Remove all marked entities Use this if you want to delete entities that were just serialized