Trait specs::saveload::Marker

pub trait Marker: Component + DeserializeOwned + Serialize + Copy {
    type Identifier: Copy + Debug + Eq + Hash;
    type Allocator: MarkerAllocator<Self>;
    fn id(&self) -> Self::Identifier;

    fn update(&mut self, update: Self) { ... }
}

This trait should be implemented by a component which is going to be used as marker. This marker should be set to entity that should be serialized. If serialization strategy needs to set marker to some entity then it should use newly allocated marker from Marker::Allocator.

Example

extern crate specs;
#[macro_use] extern crate serde;
use std::collections::HashMap;
use std::ops::Range;
use specs::{Component, Entity, DenseVecStorage};
use specs::saveload::{Marker, MarkerAllocator};

// Marker for entities that should be synced over network
#[derive(Clone, Copy, Serialize, Deserialize)]
struct NetMarker {
    id: u64,
    seq: u64,
}

impl Component for NetMarker {
    type Storage = DenseVecStorage<Self>;
}

impl Marker for NetMarker {
    type Identifier = u64;
    type Allocator = NetNode;

    fn id(&self) -> u64 {
        self.id
    }

    // Updates sequence id.
    // Entities with too old sequence id get deleted.
    fn update(&mut self, update: Self) {
        assert_eq!(self.id, update.id);
        self.seq = update.seq;
    }
}

// Each client and server has one
// Contains id range and `NetMarker -> Entity` mapping
struct NetNode {
    range: Range<u64>,
    mapping: HashMap<u64, Entity>,
}

impl MarkerAllocator<NetMarker> for NetNode {
    fn allocate(&mut self, entity: Entity, id: Option<u64>) -> NetMarker {
        let id = id.unwrap_or_else(|| {
            self.range.next().expect("Id range must be virtually endless")
        });
        let marker = NetMarker {
            id: id,
            seq: 0,
        };
        self.mapping.insert(id, entity);
        marker
    }

    fn get(&self, id: u64) -> Option<Entity> {
        self.mapping.get(&id).cloned()
    }
}

fn main() {
    use specs::World;

    let mut world = World::new();
    world.register::<NetMarker>();

    let mut node = NetNode {
        range: 0..100,
        mapping: HashMap::new(),
    };

    let entity = world.create_entity().build();
    let (marker, added) = node.mark(entity, &mut world.write::<NetMarker>());
    assert!(added);
    assert_eq!(
        node.get_marked(marker.id(), &world.entities(), &mut world.write::<NetMarker>()),
        entity
    );
}

Associated Types

type Identifier: Copy + Debug + Eq + Hash

Id of the marker

type Allocator: MarkerAllocator<Self>

Allocator for this Marker

Required Methods

fn id(&self) -> Self::Identifier

Get this marker internal id

Provided Methods

fn update(&mut self, update: Self)

Update marker with new value. It must preserve the internal Identifier.

Panics

Allowed to panic if self.id() != update.id(). But usually implementer may ignore the value of the update.id() as deserialization algorithm ensures id()s match.