Trait specs::saveload::MarkerAllocator

pub trait MarkerAllocator<M: Marker>: Resource {
    fn allocate(&mut self, entity: Entity, id: Option<M::Identifier>) -> M;

    fn get(&self, id: M::Identifier) -> Option<Entity>;

    fn mark<'a>(
        &mut self, 
        entity: Entity, 
        storage: &mut WriteStorage<'a, M>
    ) -> (M, bool) { ... }

    fn get_marked<'a>(
        &mut self, 
        id: M::Identifier, 
        entities: &Entities<'a>, 
        storage: &mut WriteStorage<'a, M>
    ) -> Entity { ... }

    fn maintain<'a>(
        &mut self, 
        _entities: &Entities<'a>, 
        _storage: &ReadStorage<'a, M>
    ) { ... }
}

This allocator is used with the Marker trait. It provides method for allocation new Markers. It should also provide a Marker -> Entity mapping. The maintain method can be implemented for cleanup and actualization. See docs for Marker for an example.

Required Methods

fn allocate(&mut self, entity: Entity, id: Option<M::Identifier>) -> M

Allocate new Marker. Stores mapping Marker -> Entity. If id argument is Some(id) then the new marker will have this id. Otherwise allocator creates marker with a new unique id.

fn get(&self, id: M::Identifier) -> Option<Entity>

Get Entity by Marker::Identifier

Provided Methods

fn mark<'a>(
&mut self,
entity: Entity,
storage: &mut WriteStorage<'a, M>
) -> (M, bool)

Create new unique marker M and attach it to entity. Or get old marker if this entity is already marked.

fn get_marked<'a>(
&mut self,
id: M::Identifier,
entities: &Entities<'a>,
storage: &mut WriteStorage<'a, M>
) -> Entity

Find an Entity by a Marker with same id and update Marker attached to the instance. Or create new entity and mark it.

fn maintain<'a>(
&mut self,
_entities: &Entities<'a>,
_storage: &ReadStorage<'a, M>
)

Maintain internal data. Cleanup if necessary.