Trait specs::saveload::SaveLoadComponent

pub trait SaveLoadComponent<M>: Component {
    type Data: Serialize + DeserializeOwned;
    type Error: Error;
    fn save<F>(&self, ids: F) -> Result<Self::Data, Self::Error>
    where
        F: FnMut(Entity) -> Option<M>;

    fn load<F>(data: Self::Data, ids: F) -> Result<Self, Self::Error>
    where
        F: FnMut(M) -> Option<Entity>;
}

This trait should be implemented in order to allow component to be serializable with SerializeSystem. It is automatically implemented for all Component + DeserializeOwned + Serialize + Copy

Associated Types

type Data: Serialize + DeserializeOwned

Serializable data representation for component

type Error: Error

Error may occur during serialization or deserialization of component

Required Methods

fn save<F>(&self, ids: F) -> Result<Self::Data, Self::Error> where
F: FnMut(Entity) -> Option<M>,

Convert this component into serializable form (Data) using entity to marker mapping function

fn load<F>(data: Self::Data, ids: F) -> Result<Self, Self::Error> where
F: FnMut(M) -> Option<Entity>,

Convert this component into deserializable form (Data) using marker to entity mapping function