Trait specs::storage::GenericWriteStorage

pub trait GenericWriteStorage {
    type Component: Component;
    fn get_mut(&mut self, entity: Entity) -> Option<&mut Self::Component>;

    fn get_mut_or_default(
        &mut self, 
        entity: Entity
    ) -> Option<&mut Self::Component>
    where
        Self::Component: Default;

    fn insert(
        &mut self, 
        entity: Entity, 
        comp: Self::Component
    ) -> InsertResult<Self::Component>;

    fn remove(&mut self, entity: Entity);

    fn _private() -> Seal;
}

Provides generic write access to WriteStorage, both as a value and a mutable reference.

Associated Types

type Component: Component

The component type of the storage

Loading content...
Required methods

fn get_mut(&mut self, entity: Entity) -> Option<&mut Self::Component>

Get mutable access to an Entitys component

fn get_mut_or_default(&mut self, entity: Entity) -> Option<&mut Self::Component> where
Self::Component: Default,

Get mutable access to an Entitys component. If the component does not exist, it is automatically created using Default::default().

Returns None if the entity is dead.

fn insert(
&mut self,
entity: Entity,
comp: Self::Component
) -> InsertResult<Self::Component>

Insert a component for an Entity

fn remove(&mut self, entity: Entity)

Remove the component for an Entity

fn _private() -> Seal

Private function to seal the trait

Loading content...