Trait specs::storage::UnprotectedStorage

pub trait UnprotectedStorage<T>: TryDefault {
    unsafe fn clean<B>(&mut self, has: B)
    where
        B: BitSetLike;

    unsafe fn get(&self, id: Index) -> &T;

    unsafe fn get_mut(&mut self, id: Index) -> &mut T;

    unsafe fn insert(&mut self, id: Index, value: T);

    unsafe fn remove(&mut self, id: Index) -> T;

    unsafe fn drop(&mut self, id: Index) { ... }
}

Used by the framework to quickly join components.

Required methods

unsafe fn clean<B>(&mut self, has: B) where
B: BitSetLike,

Clean the storage given a bitset with bits set for valid indices. Allows us to safely drop the storage.

Safety

May only be called with the mask which keeps track of the elements existing in this storage.

unsafe fn get(&self, id: Index) -> &T

Tries reading the data associated with an Index. This is unsafe because the external set used to protect this storage is absent.

Safety

May only be called after a call to insert with id and no following call to remove with id.

A mask should keep track of those states, and an id being contained in the tracking mask is sufficient to call this method.

unsafe fn get_mut(&mut self, id: Index) -> &mut T

Tries mutating the data associated with an Index. This is unsafe because the external set used to protect this storage is absent.

Safety

May only be called after a call to insert with id and no following call to remove with id.

A mask should keep track of those states, and an id being contained in the tracking mask is sufficient to call this method.

unsafe fn insert(&mut self, id: Index, value: T)

Inserts new data for a given Index.

Safety

May only be called if insert was not called with id before, or was reverted by a call to remove with `id.

A mask should keep track of those states, and an id missing from the mask is sufficient to call insert.

unsafe fn remove(&mut self, id: Index) -> T

Removes the data associated with an Index.

Safety

May only be called if an element with id was inserted and not yet removed / dropped.

Loading content...
Provided methods

unsafe fn drop(&mut self, id: Index)

Drops the data associated with an Index. This is simply more efficient than remove and can be used if the data is no longer needed.

Safety

May only be called if an element with id was inserted and not yet removed / dropped.

Loading content...