Struct specs::AsyncDispatcher

pub struct AsyncDispatcher<'a> { /* fields omitted */ }

Like, Dispatcher but works asynchronously.

Methods

impl<'a> AsyncDispatcher<'a>

fn dispatch(&mut self)

Dispatches the systems asynchronously. Does not execute thread local systems.

If you want to wait for the systems to finish, call wait().

fn wait(&mut self)

Waits for all the asynchronously dispatched systems to finish and executes thread local systems (if there are any).

fn wait_without_tl(&mut self)

Waits for all the asynchronously dispatched systems to finish without executing thread local systems.

fn is_running(&self) -> bool

Checks if any of the asynchronously dispatched systems are running.

fn dispatch_thread_local(&mut self)

Dispatch only thread local systems sequentially.

If wait_without_tl() or wait() wasn't called before, this method will wait.

fn mut_res(&mut self) -> &mut Resources

Returns the resources.

If wait_without_tl() or wait() wasn't called before, this method will do that.