Struct specs::Dispatcher

pub struct Dispatcher<'a, 'b> { /* fields omitted */ }

The dispatcher struct, allowing systems to be executed in parallel.

Methods

impl<'a, 'b> Dispatcher<'a, 'b>

fn dispatch(&mut self, res: &mut Resources)

Dispatch all the systems with given resources and context and then run thread local systems.

This function automatically redirects to

dispatch_par in case it is supported

dispatch_seq otherwise

and runs dispatch_thread_local afterwards.

fn dispatch_par(&mut self, res: &mut Resources)

Dispatches the systems (except thread local systems) in parallel given the resources to operate on.

This operation blocks the executing thread.

Only available on platforms with multithreading support (so not on emscripten).

fn dispatch_seq(&mut self, res: &mut Resources)

Dispatches the systems (except thread local systems) sequentially.

This is useful if parallel overhead is too big or the platform does not support multithreading.

fn dispatch_thread_local(&mut self, res: &mut Resources)

Dispatch only thread local systems sequentially.