Struct specs::Entities

pub struct Entities<'a> { /* fields omitted */ }

A custom entity guard used to hide the the fact that Generations is lazily created and updated. For this to be useful it must be joined with a component. This is because the Generation table includes every possible Generation of Entities even if they have never existed.