Struct specs::FetchArg

pub struct FetchArg<'a>(_);

System fetch-time argument. The fetch is executed at the start of the run. It contains a subset of World methods that make sense during initialization.

Methods

impl<'a> FetchArg<'a>

fn read<T: Component>(&self) -> RwLockReadGuard<'a, T::Storage>

Lock a component for reading.

fn write<T: Component>(&self) -> RwLockWriteGuard<'a, T::Storage>

Lock a component for writing.

fn entities(self) -> EntityIter<'a>

Return the entity iterator.