Struct specs::MaskedStorage

pub struct MaskedStorage<T: Component> { /* fields omitted */ }

The UnprotectedStorage together with the BitSet that knows about which elements are stored, and which are not.

Methods

impl<T: Component> MaskedStorage<T>

pub fn new() -> MaskedStorage<T>

Creates a new MaskedStorage. This is called when you register a new component type within the world.

pub fn clear(&mut self)

Clear the contents of this storage.

pub fn remove(&mut self, id: Index) -> Option<T>

Remove an element by a given index.