Struct specs::OccupiedEntry

pub struct OccupiedEntry<'a, 'b: 'a, T: 'a, D: 'a> { /* fields omitted */ }

An entry to a storage which has a component associated to the entity.

Methods

impl<'a, 'b, T, D> OccupiedEntry<'a, 'b, T, D> where
T: Component,
D: Deref<Target = MaskedStorage<T>>,

pub fn get(&self) -> &T

Get a reference to the component associated with the entity.

impl<'a, 'b, T, D> OccupiedEntry<'a, 'b, T, D> where
T: Component,
D: DerefMut<Target = MaskedStorage<T>>,

pub fn get_mut(&mut self) -> &mut T

Get a mutable reference to the component associated with the entity.

pub fn into_mut(self) -> &'a mut T

Converts the OccupiedEntry into a mutable reference bounded by the storage's lifetime.

pub fn insert(&mut self, component: T) -> T

Inserts a value into the storage and returns the old one.

pub fn remove(self) -> T

Removes the component from the storage and returns it.