Struct specs::RestrictedStorage

pub struct RestrictedStorage<'rf, 'st: 'rf, B, T, R, RT> where
    T: Component,
    R: Borrow<T::Storage> + 'rf,
    B: Borrow<BitSet> + 'rf,  { /* fields omitted */ }

Similar to a MaskedStorage and a Storage combined, but restricts usage to only getting and modifying the components. That means nothing that would modify the inner bitset so the iteration cannot be invalidated. For example, no insertion or removal is allowed.

Example Usage:

struct SomeComp(u32);
impl Component for SomeComp {
    type Storage = VecStorage<Self>;
}

struct RestrictedSystem;
impl<'a> System<'a> for RestrictedSystem {
    type SystemData = (
        Entities<'a>,
        WriteStorage<'a, SomeComp>,
    );
    fn run(&mut self, (entities, mut some_comps): Self::SystemData) {
        for (entity, (mut entry, restricted)) in (
            &*entities,
            &mut some_comps.restrict()
        ).join() {
            // Check if the reference is fine to mutate.
            if restricted.get_unchecked(&entry).0 < 5 {
                // Get a mutable reference now.
                let mut mutable = restricted.get_mut_unchecked(&mut entry);
                mutable.0 += 1;
            }
        }
    }
}

Methods

impl<'rf, 'st, B, T, R, RT> RestrictedStorage<'rf, 'st, B, T, R, RT> where
T: Component,
R: Borrow<T::Storage>,
B: Borrow<BitSet>,

pub fn get(&self, entity: Entity) -> Option<&T>

Attempts to get the component related to the entity.

Functions similar to the normal Storage::get implementation.

pub fn get_unchecked(&self, entry: &Entry<'rf, T>) -> &T

Gets the component related to the current entry without checking whether the storage has it or not.

impl<'rf, 'st, B, T, R, RT> RestrictedStorage<'rf, 'st, B, T, R, RT> where
T: Component,
R: BorrowMut<T::Storage>,
B: Borrow<BitSet>,

pub fn get_mut_unchecked(&mut self, entry: &Entry<'rf, T>) -> &mut T

Gets the component related to the current entry without checking whether the storage has it or not.

impl<'rf, 'st, B, T, R> RestrictedStorage<'rf, 'st, B, T, R, NormalRestriction> where
T: Component,
R: BorrowMut<T::Storage>,
B: Borrow<BitSet>,

pub fn get_mut(&mut self, entity: Entity) -> Option<&mut T>

Attempts to get the component related to the entity mutably.

Functions similar to the normal Storage::get_mut implementation.

Note: This only works if this is a non-parallel RestrictedStorage.