Trait specs::UnprotectedStorage

pub trait UnprotectedStorage<T>: Default + Sized {
    unsafe fn clean<F>(&mut self, f: F)
    where
        F: Fn(Index) -> bool;

    unsafe fn get(&self, id: Index) -> &T;

    unsafe fn get_mut(&mut self, id: Index) -> &mut T;

    unsafe fn insert(&mut self, id: Index, value: T);

    unsafe fn remove(&mut self, id: Index) -> T;
}

Used by the framework to quickly join components.

Required Methods

unsafe fn clean<F>(&mut self, f: F) where
F: Fn(Index) -> bool,

Clean the storage given a check to figure out if an index is valid or not. Allows us to safely drop the storage.

unsafe fn get(&self, id: Index) -> &T

Tries reading the data associated with an Index. This is unsafe because the external set used to protect this storage is absent.

unsafe fn get_mut(&mut self, id: Index) -> &mut T

Tries mutating the data associated with an Index. This is unsafe because the external set used to protect this storage is absent.

unsafe fn insert(&mut self, id: Index, value: T)

Inserts new data for a given Index.

unsafe fn remove(&mut self, id: Index) -> T

Removes the data associated with an Index.