Crate specs

SPECS Parallel ECS

This library provides an ECS variant designed for parallel execution and convenient usage. It is highly flexible when it comes to actual component data and the way it is stored and accessed.

Features:

depending on chosen features either 0 virtual function calls or one per system

parallel iteration over components

parallel execution of systems

High-level overview

One could basically split this library up into two parts: The data part and the execution part.

The data

World is where component storages, resources and entities are stored. See the docs of World for more.

Components can be easily implemented like this:

use specs::prelude::*;

struct MyComp;

impl Component for MyComp {
    type Storage = VecStorage<MyComp>;
}

You can choose different storages according to your needs.

These storages can be joined together, for example joining a Velocity and a Position storage means you'll only get entities which have both of them. Thanks to rayon, this is even possible in parallel! See ParJoin for more.

System execution

Here we have System and Dispatcher as our core types. Both types are provided by a library called shred.

The Dispatcher can be seen as an optional part here; it allows dispatching the systems in parallel, given a list of systems and their dependencies on other systems.

If you don't like it, you can also execute the systems yourself by using RunNow.

Systems are traits with a run() method and an associated SystemData, allowing type-safe aspects (knowledge about the reads / writes of the systems).

Examples

This is a basic example of using Specs:

extern crate specs;

use specs::prelude::*;

// A component contains data
// which is associated with an entity.
struct Vel(f32);
struct Pos(f32);

impl Component for Vel {
    type Storage = VecStorage<Vel>;
}

impl Component for Pos {
    type Storage = VecStorage<Pos>;
}

struct SysA;

impl<'a> System<'a> for SysA {
    // These are the resources required for execution.
    // You can also define a struct and `#[derive(SystemData)]`,
    // see the `full` example.
    type SystemData = (WriteStorage<'a, Pos>, ReadStorage<'a, Vel>);

    fn run(&mut self, data: Self::SystemData) {
        // The `.join()` combines multiple components,
        // so we only access those entities which have
        // both of them.

        let (mut pos, vel) = data;

        // This joins the component storages for Position
        // and Velocity together; it's also possible to do this
        // in parallel using rayon's `ParallelIterator`s.
        // See `ParJoin` for more.
        for (pos, vel) in (&mut pos, &vel).join() {
            pos.0 += vel.0;
        }
    }
}

fn main() {
    // The `World` is our
    // container for components
    // and other resources.

    let mut world = World::new();
    world.register::<Pos>();
    world.register::<Vel>();

    // An entity may or may not contain some component.

    world.create_entity().with(Vel(2.0)).with(Pos(0.0)).build();
    world.create_entity().with(Vel(4.0)).with(Pos(1.6)).build();
    world.create_entity().with(Vel(1.5)).with(Pos(5.4)).build();

    // This entity does not have `Vel`, so it won't be dispatched.
    world.create_entity().with(Pos(2.0)).build();

    // This builds a dispatcher.
    // The third parameter of `add` specifies
    // logical dependencies on other systems.
    // Since we only have one, we don't depend on anything.
    // See the `full` example for dependencies.
    let mut dispatcher = DispatcherBuilder::new().add(SysA, "sys_a", &[]).build();

    // This dispatches all the systems in parallel (but blocking).
    dispatcher.dispatch(&mut world.res);
}

See the repository's examples directory for more examples.

Modules

data
Reexports for types implementing SystemData.

entity
Entity related types.

prelude
Reexports for very common types.

storages
Different types of storages you can use for your components.

Structs

AsyncDispatcher
Like, Dispatcher but works asynchronously.

CheckStorage
A storage type that iterates entities that have a particular component type, but does not return the component.

Dispatcher
The dispatcher struct, allowing systems to be executed in parallel.

DispatcherBuilder
Builder for the Dispatcher.

Entities
The entities of this ECS.

Entity
Entity type, as seen by the user.

JoinIter
JoinIter is an Iterator over a group of Storages.

JoinParIter
JoinParIter is a ParallelIterator over a group of Storages.

World
The World struct contains the component storages and other resources.

Enums

InsertResult
The status of an insert()ion into a storage.

RunningTime

Traits

Component
Abstract component type. Doesn't have to be Copy or even Clone.

Join
The purpose of the Join trait is to provide a way to access multiple storages at the same time with the merged bit set.

ParJoin
The purpose of the ParJoin trait is to provide a way to access multiple storages in parallel at the same time with the merged bit set.

Resource
A resource defines a set of data which can only be accessed according to Rust's typical borrowing model (one writer xor multiple readers).

RunNow
Trait for fetching data and running systems. Automatically implemented for systems.

System
A System, executed with a set of required Resources.

UnprotectedStorage
Used by the framework to quickly join components.

Type Definitions

Index
An index is basically the id of an Entity.