Struct specs::common::Merge

pub struct Merge<F> { /* fields omitted */ }

A system which merges Ready futures into the persistent storage. Please note that your World has to contain a component storage for F and F::Item.

In case of an error, it will be added to the Errors resource.

Methods

impl<F> Merge<F>

pub fn new() -> Self

Creates a new merge system.