Trait specs::join::ParJoin

pub unsafe trait ParJoin: Join {
    fn par_join(self) -> JoinParIter<Self>
    where
        Self: Sized,
    { ... }
}

The purpose of the ParJoin trait is to provide a way to access multiple storages in parallel at the same time with the merged bit set.

Provided Methods

fn par_join(self) -> JoinParIter<Self> where
Self: Sized,

Create a joined parallel iterator over the contents.