Trait specs::System

pub trait System<'a> {
    type SystemData: SystemData<'a>;
    fn run(&mut self, data: Self::SystemData);

    fn running_time(&self) -> RunningTime { ... }
}

A System, executed with a set of required Resources.

Associated Types

type SystemData: SystemData<'a>

The resource bundle required to execute this system.

To create such a resource bundle, simple derive SystemData for it.

Required Methods

fn run(&mut self, data: Self::SystemData)

Executes the system with the required system data.

Provided Methods

fn running_time(&self) -> RunningTime

Returns a hint how long the system needs for running. This is used to optimize the way they're executed (might allow more parallelization).

Defaults to RunningTime::Average.